// gen/typescript/Token_serde.ts
export interface Jws {
  algorithm: string;
  keyId: string | null;
  signature: Uint8Array;
}